+++ src/main/java/com/goxr3plus/streamplayer/stream/FileDataSource.java
@@ -1,12 +1,13 @@
 package com.goxr3plus.streamplayer.stream;
 
 import javax.sound.sampled.AudioFileFormat;
+import javax.sound.sampled.AudioInputStream;
 import javax.sound.sampled.AudioSystem;
 import javax.sound.sampled.UnsupportedAudioFileException;
 import java.io.File;
 import java.io.IOException;
 
-public class FileDataSource extends DataSource {
+public class FileDataSource extends DataSourceBase implements DataSource {
 
     File source;
 
@@ -18,5 +19,10 @@ public class FileDataSource extends DataSource {
     @Override
     public AudioFileFormat getAudioFileFormat() throws UnsupportedAudioFileException, IOException {
         return AudioSystem.getAudioFileFormat(this.source);
+    }
+
+    @Override
+    public AudioInputStream getAudioInputStream() throws UnsupportedAudioFileException, IOException {
+        return AudioSystem.getAudioInputStream(source);
     }
 }
