+++ src/main/java/com/goxr3plus/streamplayer/stream/StreamDataSource.java
@@ -1,12 +1,13 @@
 package com.goxr3plus.streamplayer.stream;
 
 import javax.sound.sampled.AudioFileFormat;
+import javax.sound.sampled.AudioInputStream;
 import javax.sound.sampled.AudioSystem;
 import javax.sound.sampled.UnsupportedAudioFileException;
 import java.io.IOException;
 import java.io.InputStream;
 
-public class StreamDataSource extends DataSource {
+public class StreamDataSource extends DataSourceBase implements DataSource {
 
     InputStream source;
 
@@ -18,5 +19,10 @@ public class StreamDataSource extends DataSource {
     @Override
     public AudioFileFormat getAudioFileFormat() throws UnsupportedAudioFileException, IOException {
         return AudioSystem.getAudioFileFormat(source);
+    }
+
+    @Override
+    public AudioInputStream getAudioInputStream() throws UnsupportedAudioFileException, IOException {
+        return AudioSystem.getAudioInputStream(source);
     }
 }
