+++ src/main/java/com/goxr3plus/streamplayer/stream/UrlDataSource.java
@@ -1,12 +1,13 @@
 package com.goxr3plus.streamplayer.stream;
 
 import javax.sound.sampled.AudioFileFormat;
+import javax.sound.sampled.AudioInputStream;
 import javax.sound.sampled.AudioSystem;
 import javax.sound.sampled.UnsupportedAudioFileException;
 import java.io.IOException;
 import java.net.URL;
 
-public class UrlDataSource extends DataSource {
+public class UrlDataSource extends DataSourceBase implements DataSource {
 
     URL source;
 
@@ -18,5 +19,10 @@ public class UrlDataSource extends DataSource {
     @Override
     public AudioFileFormat getAudioFileFormat() throws UnsupportedAudioFileException, IOException {
         return AudioSystem.getAudioFileFormat((URL) source);
+    }
+
+    @Override
+    public AudioInputStream getAudioInputStream() throws UnsupportedAudioFileException, IOException {
+        return AudioSystem.getAudioInputStream(source);
     }
 }
